+++ es4x/src/main/java/io/reactiverse/es4x/ECMAEngine.java
@@ -16,7 +16,6 @@
 package io.reactiverse.es4x;
 
 import io.reactiverse.es4x.impl.graal.GraalEngine;
-import io.reactiverse.es4x.impl.nashorn.NashornEngine;
 import io.vertx.core.Vertx;
 
 import java.util.regex.Pattern;
@@ -24,31 +23,7 @@ import java.util.regex.Pattern;
 public interface ECMAEngine {
 
   static ECMAEngine newEngine(Vertx vertx) {
-    String userSelectedEngine = System.getProperty("es4x.engine", "graaljs");
-
-    if ("graaljs".equals(userSelectedEngine)) {
-      try {
-
-        System.setProperty("es4x.engine", "graaljs");
-        return new GraalEngine(vertx);
-      } catch (NoClassDefFoundError | IllegalStateException e0) {
-        // in the case classes are missing, the graal bits are missing
-        // so fallback to Nashorn
-
-        // we could also have an illegal state when the graal is missing
-        // the language bits, in that case also try to fallback to nashorn
-
-        // force the engine to nashorn
-        userSelectedEngine = "nashorn";
-      }
-    }
-
-    if ("nashorn".equals(userSelectedEngine)) {
-      System.setProperty("es4x.engine", "nashorn");
-      return new NashornEngine(vertx);
-    }
-
-    throw new IllegalStateException("Unsupported runtime [" + userSelectedEngine + "]");
+    return new GraalEngine(vertx);
   }
 
   default Pattern[] allowedHostClassFilters() {