+++ es4x/src/test/java/io/reactiverse/es4x/test/FactoryMJSTest.java
@@ -1,43 +1,16 @@
 package io.reactiverse.es4x.test;
 
-import io.reactiverse.es4x.Runtime;
 import io.vertx.core.DeploymentOptions;
 import io.vertx.ext.unit.Async;
 import io.vertx.ext.unit.TestContext;
 import io.vertx.ext.unit.junit.RunTestOnContext;
-import io.vertx.ext.unit.junit.VertxUnitRunnerWithParametersFactory;
-import org.junit.Before;
+import io.vertx.ext.unit.junit.VertxUnitRunner;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
 
-import java.util.Arrays;
-import java.util.List;
-
-import static io.reactiverse.es4x.test.Helper.getRuntime;
-
-@RunWith(Parameterized.class)
-@Parameterized.UseParametersRunnerFactory(VertxUnitRunnerWithParametersFactory.class)
+@RunWith(VertxUnitRunner.class)
 public class FactoryMJSTest {
-
-  @Parameterized.Parameters
-  public static List<String> engines() {
-    return Arrays.asList(/*"Nashorn", */"GraalJS");
-  }
-
-  private final String engineName;
-  private Runtime runtime;
-
-  public FactoryMJSTest(String engine) {
-    System.setProperty("es4x.engine", engine.toLowerCase());
-    engineName = engine;
-  }
-
-  @Before
-  public void initialize() {
-    runtime = getRuntime(rule.vertx(), engineName);
-  }
 
   @Rule
   public RunTestOnContext rule = new RunTestOnContext();