+++ es4x/src/test/java/io/reactiverse/es4x/test/FactoryTest.java
@@ -3,27 +3,13 @@ package io.reactiverse.es4x.test;
 import io.vertx.ext.unit.Async;
 import io.vertx.ext.unit.TestContext;
 import io.vertx.ext.unit.junit.RunTestOnContext;
-import io.vertx.ext.unit.junit.VertxUnitRunnerWithParametersFactory;
+import io.vertx.ext.unit.junit.VertxUnitRunner;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
 
-import java.util.Arrays;
-import java.util.List;
-
-@RunWith(Parameterized.class)
-@Parameterized.UseParametersRunnerFactory(VertxUnitRunnerWithParametersFactory.class)
+@RunWith(VertxUnitRunner.class)
 public class FactoryTest {
-
-  @Parameterized.Parameters
-  public static List<String> engines() {
-    return Arrays.asList("Nashorn", "GraalJS");
-  }
-
-  public FactoryTest(String engine) {
-    System.setProperty("es4x.engine", engine.toLowerCase());
-  }
 
   @Rule
   public RunTestOnContext rule = new RunTestOnContext();