+++ es4x/src/test/java/io/reactiverse/es4x/test/FilterHostClassTest.java
@@ -1,32 +1,32 @@
 package io.reactiverse.es4x.test;
 
+import io.vertx.core.logging.Logger;
+import io.vertx.core.logging.LoggerFactory;
 import io.vertx.ext.unit.Async;
 import io.vertx.ext.unit.TestContext;
 import io.vertx.ext.unit.junit.RunTestOnContext;
-import io.vertx.ext.unit.junit.VertxUnitRunnerWithParametersFactory;
+import io.vertx.ext.unit.junit.VertxUnitRunner;
+import org.junit.AfterClass;
+import org.junit.BeforeClass;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
 
-import javax.swing.text.html.parser.Parser;
-import java.util.Arrays;
-import java.util.List;
-import java.util.regex.Pattern;
-
-@RunWith(Parameterized.class)
-@Parameterized.UseParametersRunnerFactory(VertxUnitRunnerWithParametersFactory.class)
+@RunWith(VertxUnitRunner.class)
 public class FilterHostClassTest {
 
-  @Parameterized.Parameters
-  public static List<String> engines() {
-    return Arrays.asList("Nashorn", "GraalJS");
-  }
+  private static final Logger LOGGER = LoggerFactory.getLogger(FilterHostClassTest.class);
 
-  public FilterHostClassTest(String engine) {
-    System.setProperty("es4x.engine", engine.toLowerCase());
+
+  @BeforeClass
+  public static void before() {
     //System.setProperty("es4x.host.class.filter", "io.vertx.**,java.util.**,java.time.**,java.lang.**,java.net.**,io.reactiverse.es4x.**,!java.nio.file.FileAlreadyExistsException");
     System.setProperty("es4x.host.class.filter", "!java.nio.file.FileAlreadyExistsException");
+  }
+
+  @AfterClass
+  public static void after() {
+    System.getProperties().remove("es4x.host.class.filter");
   }
 
   @Rule
@@ -37,7 +37,7 @@ public class FilterHostClassTest {
     final Async async = ctx.async();
     rule.vertx().deployVerticle("js:./verticle3.js", deploy -> {
       ctx.assertFalse(deploy.succeeded());
-      System.out.println(deploy.cause());
+      LOGGER.error("SHOULD FAIL", deploy.cause());
       // fails because we excluded the class "java.nio.file.FileAlreadyExistsException"
       async.complete();
     });