+++ es4x/src/test/java/io/reactiverse/es4x/test/LoggerTest.java
@@ -1,43 +1,29 @@
 package io.reactiverse.es4x.test;
 
 import io.reactiverse.es4x.Runtime;
+import io.reactiverse.es4x.impl.graal.GraalEngine;
 import io.vertx.core.logging.Logger;
 import io.vertx.core.logging.LoggerFactory;
 import io.vertx.ext.unit.junit.RunTestOnContext;
+import io.vertx.ext.unit.junit.VertxUnitRunner;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
 
-import java.util.Arrays;
-import java.util.List;
-
-import static io.reactiverse.es4x.test.Helper.getRuntime;
-
-@RunWith(Parameterized.class)
+@RunWith(VertxUnitRunner.class)
 public class LoggerTest {
 
   private final Logger log = LoggerFactory.getLogger(LoggerTest.class);
 
-  @Parameterized.Parameters
-  public static List<String> engines() {
-    return Arrays.asList("Nashorn", "GraalJS");
-  }
-
-  private final String engineName;
   private Runtime runtime;
-
-  public LoggerTest(String engine) {
-    engineName = engine;
-  }
 
   @Rule
   public RunTestOnContext rule = new RunTestOnContext();
 
   @Before
   public void initialize() {
-    runtime = getRuntime(rule.vertx(), engineName);
+    runtime = new GraalEngine(rule.vertx()).newContext();
   }
 
   @Test