+++ es4x/src/test/java/io/reactiverse/es4x/test/ProcessTest.java
@@ -1,43 +1,31 @@
 package io.reactiverse.es4x.test;
 
 import io.reactiverse.es4x.Runtime;
+import io.reactiverse.es4x.impl.graal.GraalEngine;
 import io.vertx.ext.unit.junit.RunTestOnContext;
+import io.vertx.ext.unit.junit.VertxUnitRunner;
 import org.graalvm.polyglot.Value;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
 
-import java.util.Arrays;
-import java.util.List;
 import java.util.Map;
 
-import static io.reactiverse.es4x.test.Helper.getRuntime;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.fail;
 
-@RunWith(Parameterized.class)
+@RunWith(VertxUnitRunner.class)
 public class ProcessTest {
 
-  @Parameterized.Parameters
-  public static List<String> engines() {
-    return Arrays.asList("Nashorn", "GraalJS");
-  }
-
-  private final String engineName;
   private Runtime runtime;
-
-  public ProcessTest(String engine) {
-    engineName = engine;
-  }
 
   @Rule
   public RunTestOnContext rule = new RunTestOnContext();
 
   @Before
   public void initialize() {
-    runtime = getRuntime(rule.vertx(), engineName);
+    runtime = new GraalEngine(rule.vertx()).newContext();
   }
 
   @Test(timeout = 10000)