+++ es4x/src/test/resources/cluster/receiver.js
@@ -2,7 +2,7 @@ var eb = vertx.eventBus();
 
 eb.consumer("ping-address", function (message) {
 
-  console.log("Received message: " + message.body());
+  console.log("Receiver received: " + message.body());
   // Now send back reply
   message.reply([1,2,3]);
 });